pdfproject/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.0
cssselect2==0.7.0
Django==5.1.1
djangorestframework==3.15.2
fonttools==4.53.1
gunicorn==23.0.0
html5lib==1.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
tzdata==2024.1
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
